// frontend/types/api.ts
export interface components {
    schemas: {
        /** HTTPValidationError */
        HTTPValidationError: {
            /** Detail */
            detail?: components["schemas"]["ValidationError"][];
        };
        /**
         * PriorityEnum
         * @enum {string}
         */
        PriorityEnum: "low" | "medium" | "high";
        /** TodoCreate */
        TodoCreate: {
            /** Title */
            title: string;
            /** Description */
            description?: string | null;
            /**
             * Completed
             * @default false
             */
            completed: boolean;
            priority?: components["schemas"]["PriorityEnum"] | null;
            /** Due Date */
            due_date?: string | null;
        };
        /** TodoRead */
        TodoRead: {
            /** Title */
            title: string;
            /** Description */
            description?: string | null;
            /**
             * Completed
             * @default false
             */
            completed: boolean;
            priority?: components["schemas"]["PriorityEnum"] | null;
            /** Due Date */
            due_date?: string | null;
            /** Id */
            id: number;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /**
             * Updated At
             * Format: date-time
             */
            updated_at: string;
        };
        /** TodoUpdate */
        TodoUpdate: {
            /** Title */
            title?: string | null;
            /** Description */
            description?: string | null;
            /** Completed */
            completed?: boolean | null;
            priority?: components["schemas"]["PriorityEnum"] | null;
            /** Due Date */
            due_date?: string | null;
        };
        /** ValidationError */
        ValidationError: {
            /** Location */
            loc: (string | number)[];
            /** Message */
            msg: string;
            /** Error Type */
            type: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}